evaluation/requirements.txt
moverscore
bert-score
nltk==3.5